# go/store/skip: speedup skip list