… go/store/skip/list_test.go