micro benchmark results
```
goos: darwin
goarch: amd64
pkg: github.com/dolthub/dolt/go/store/skip
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
                                   │    before     │                after                │
                                   │    sec/op     │   sec/op     vs base                │
Get/unsorted_keys/n=64-12            182.65n ± 12%   90.91n ± 5%  -50.23% (p=0.000 n=10)
Get/unsorted_keys/n=2048-12           334.8n ±  7%   186.5n ± 4%  -44.30% (p=0.000 n=10)
Get/unsorted_keys/n=65536-12          670.1n ±  5%   475.7n ± 4%  -29.02% (p=0.000 n=10)
Get/ascending_keys/n=64-12           188.75n ± 11%   88.27n ± 7%  -53.24% (p=0.000 n=10)
Get/ascending_keys/n=2048-12          318.8n ±  3%   156.1n ± 7%  -51.03% (p=0.000 n=10)
Get/ascending_keys/n=65536-12         470.0n ±  8%   218.7n ± 5%  -53.47% (p=0.000 n=10)
Get/descending_keys/n=64-12          214.90n ± 11%   90.30n ± 5%  -57.98% (p=0.000 n=10)
Get/descending_keys/n=2048-12         305.3n ±  7%   151.1n ± 3%  -50.50% (p=0.000 n=10)
Get/descending_keys/n=65536-12        467.3n ±  9%   217.5n ± 6%  -53.46% (p=0.000 n=10)
Put/unsorted_keys/n=64-12             263.2n ± 11%   106.6n ± 6%  -59.53% (p=0.000 n=10)
Put/unsorted_keys/n=2048-12           397.1n ±  8%   206.0n ± 2%  -48.12% (p=0.000 n=10)
Put/unsorted_keys/n=65536-12          686.8n ±  5%   449.4n ± 2%  -34.55% (p=0.000 n=10)
Put/asending_keys/n=64-12            232.15n ± 41%   89.07n ± 6%  -61.63% (p=0.000 n=10)
Put/asending_keys/n=2048-12           365.2n ± 13%   143.1n ± 9%  -60.82% (p=0.000 n=10)
Put/asending_keys/n=65536-12          451.4n ±  9%   206.5n ± 7%  -54.27% (p=0.000 n=10)
Put/descending_keys/n=64-12          188.25n ± 14%   71.79n ± 4%  -61.86% (p=0.000 n=10)
Put/descending_keys/n=2048-12        209.30n ± 27%   83.79n ± 3%  -59.97% (p=0.000 n=10)
Put/descending_keys/n=65536-12        229.2n ±  9%   102.2n ± 1%  -55.38% (p=0.000 n=10)
IterAll/unsorted_keys/n=64-12         469.6n ± 10%   166.2n ± 1%  -64.61% (p=0.000 n=10)
IterAll/unsorted_keys/n=2048-12      18.622µ ±  8%   9.711µ ± 1%  -47.85% (p=0.000 n=10)
IterAll/unsorted_keys/n=65536-12     1214.1µ ±  5%   950.5µ ± 4%  -21.71% (p=0.000 n=10)
IterAll/asending_keys/n=64-12         456.8n ± 22%   167.4n ± 1%  -63.36% (p=0.000 n=10)
IterAll/asending_keys/n=2048-12      16.380µ ± 10%   7.829µ ± 8%  -52.20% (p=0.000 n=10)
IterAll/asending_keys/n=65536-12      598.6µ ±  8%   313.7µ ± 4%  -47.60% (p=0.000 n=10)
IterAll/descending_keys/n=64-12       483.3n ± 28%   166.3n ± 1%  -65.58% (p=0.000 n=10)
IterAll/descending_keys/n=2048-12    17.414µ ± 16%   7.964µ ± 4%  -54.27% (p=0.000 n=10)
IterAll/descending_keys/n=65536-12    621.1µ ± 12%   338.9µ ± 3%  -45.45% (p=0.000 n=10)
geomean                                                           -52.90%

```